requirements.txt
opendatasets==0.1.22
faiss-cpu==1.12.0
sentence-transformers==5.1.1
langchain-community==0.3.30
PyMuPDF==1.26.4
python-dotenv==1.1.1
fastapi==0.118.0
uvicorn==0.37.0
numpy==2.3.3 
pandas==2.3.3 